feat: add initial hero section version
