+++ src/pages/index.astro
@@ -9,6 +9,7 @@ import CrewJoinComp from "../components/sections/CrewJoinComp.astro";
 import TechBlogComp from "../components/sections/TechBlogComp.astro";
 import TicketsComp from "../components/sections/TicketsComp.astro";
 import ExpoComp from "../components/sections/ExpoComp.astro";
+import Hero from "../components/Hero.astro";
 ---
 
 <Layout title="The Gathering">
@@ -17,18 +18,21 @@ import ExpoComp from "../components/sections/ExpoComp.astro";
       <H1 text="BLI MED I NORGES STØRSTE PÅSKEHYTTE" />
     </div>
     <div class="grid sm:grid-cols-5 gap-4 w-full">
-      <div class="sm:col-start-4 sm:col-span-2 sm:row-start-1">
+      <div class="sm:col-span-5">
+        <Hero image="/images/tg19-lights.jpg" />
+      </div>
+      <div class="sm:col-start-4 sm:col-span-2 sm:row-start-2">
         <TicketsComp />
       </div>
       <div class="sm:col-span-5 my-2">
         <NewsComp />
       </div>
       <div
-        class="sm:col-start-1 sm:col-span-3 sm:row-start-1 sm:row-span-2 flex"
+        class="sm:col-start-1 sm:col-span-3 sm:row-start-2 sm:row-span-2 flex"
       >
         <AboutComp />
       </div>
-      <div class="sm:col-start-4 sm:col-span-2 sm:row-start-2">
+      <div class="sm:col-start-4 sm:col-span-2 sm:row-start-3">
         <ExpoComp />
       </div>
       <div class="sm:col-span-2">
