+++ src/utils/articles.ts
@@ -1,62 +1,4 @@
-export interface Article {
-  id: number;
-  meta: {
-    type: string;
-    detail_url: string;
-    html_url: string;
-    slug: string;
-    first_published_at: string;
-    seo_title: string;
-    search_description: string;
-    locale?: "en" | "no";
-  };
-  title: string;
-  intro: string;
-  body: string;
-  contributors: Array<{
-    id: number;
-    name: string;
-    contribution_type: string;
-    image?: string;
-  }>;
-  tags: Array<{
-    id: number;
-    name: string;
-    slug: string;
-  }>;
-  main_image?: {
-    title: string;
-    alt: string;
-    url: string;
-    sizes: {
-      thumbnail: {
-        url: string;
-        width: number;
-        height: number;
-      };
-      small: {
-        url: string;
-        width: number;
-        height: number;
-      };
-      medium: {
-        url: string;
-        width: number;
-        height: number;
-      };
-      large: {
-        url: string;
-        width: number;
-        height: number;
-      };
-      extra_large: {
-        url: string;
-        width: number;
-        height: number;
-      };
-    };
-  };
-}
+import type { Article } from "../types";
 
 export interface FetchArticlesProps {
   api_url: string;
